refactor(shared): extract legalPause module
Moves calculateLegalPauseStatus, calculateLegalPauseZones and the internal
buildPauseTiers helper into legalPause.ts. They form a self-contained tier
model and were the largest cohesive block in index.ts.
